+++ src/PAL/COM/sockets/ssl/mbedTLS/ssl_initialize_internal.cpp
@@ -1,0 +1,14 @@
+//
+// Copyright (c) 2018 The nanoFramework project contributors
+// Portions Copyright (c) 2006-2015, ARM Limited, All Rights Reserved
+// Portions Copyright (c) Microsoft Corporation.  All rights reserved.
+// See LICENSE file in the project root for full license information.
+//
+
+#include <ssl.h> 
+#include "mbedtls.h"
+
+bool ssl_initialize_internal()
+{
+     return true;
+}
