+++ src/PAL/COM/sockets/ssl/mbedTLS/ssl_parse_certificate_internal.cpp
